Modified MetaTag model of base app. Modified BaseView of base app. Added templates common_footer.html and destination_slider into base app's template. Updated base, blog_slider and footer templates of base app. Replaced content field with rich text editor field. Replaced meta tags field with many to many field. Added property get_meta_tags. Added social media fields into company model of company app. Modified CompanyBaseView of company app. Provided a slug generation functionality for TouristDestination model. Modified templates, model, urls and views of educational app.
Added a preview method for url type in creation and updation forms of all multipages.
Provided a default meta title if it is not provided.
Added Modified tag selection with creation in multipages.
Provided a functionality to hide support languages of detail pages.
Provided a text editor for meta tag creation and updation.